+++ EnemizerLibrary/Dungeons/Dungeon.cs
@@ -6,24 +6,6 @@ using System.Threading.Tasks;
 
 namespace EnemizerLibrary
 {
-    public enum DungeonType
-    {
-        EasternPalace,
-        DesertPalace,
-        TowerOfHera,
-        PalaceOfDarkness,
-        SwampPalace,
-        SkullWoods,
-        ThievesTown,
-        IcePalace,
-        MiseryMire,
-        TurtleRock,
-        GanonsTower1,
-        GanonsTower2,
-        GanonsTower3,
-        NotSet
-    }
-
     public abstract class Dungeon
     {
         public string Name { get; set; }
@@ -46,210 +28,4 @@ namespace EnemizerLibrary
         }
     }
 
-    public class EasternPalaceDungeon : Dungeon
-    {
-        public EasternPalaceDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Eastern Palace";
-            DungeonType = DungeonType.EasternPalace;
-            DungeonCrystalTypeAddress = CrystalConstants.EasternPalaceCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.EasternPalaceCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 200;
-            BossAddress = 0x04D7BE;
-            BossDropItemAddress = 0x180150;
-        }
-    }
-
-    public class DesertPalaceDungeon : Dungeon
-    {
-        public DesertPalaceDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Desert Palace";
-            DungeonType = DungeonType.DesertPalace;
-            DungeonCrystalTypeAddress = CrystalConstants.DesertPalaceCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.DesertPalaceCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 51;
-            BossAddress = 0x04D694;
-            BossDropItemAddress = 0x180151;
-        }
-    }
-
-    public class TowerOfHeraDungeon : Dungeon
-    {
-        public TowerOfHeraDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Tower of Hera";
-            DungeonType = DungeonType.TowerOfHera;
-            DungeonCrystalTypeAddress = CrystalConstants.TowerOfHeraCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.TowerOfHeraCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 7;
-            BossAddress = 0x04D63C;
-            BossDropItemAddress = 0x180152;
-
-            DisallowedBosses.Add(BossType.Armos);
-            DisallowedBosses.Add(BossType.Arrghus);
-            DisallowedBosses.Add(BossType.Blind);
-            DisallowedBosses.Add(BossType.Lanmola);
-            DisallowedBosses.Add(BossType.Trixnexx);
-        }
-    }
-
-    public class PalaceOfDarknessDungeon : Dungeon
-    {
-        public PalaceOfDarknessDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Palace of Darkness";
-            DungeonType = DungeonType.PalaceOfDarkness;
-            DungeonCrystalTypeAddress = CrystalConstants.PalaceOfDarknessCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.PalaceOfDarknessCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 90;
-            BossAddress = 0x04D6E2;
-            BossDropItemAddress = 0x180153;
-        }
-    }
-
-    public class SwampPalaceDungeon : Dungeon
-    {
-        public SwampPalaceDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Swamp Palace";
-            DungeonType = DungeonType.SwampPalace;
-            DungeonCrystalTypeAddress = CrystalConstants.SwampPalaceCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.SwampPalaceCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 6;
-            BossAddress = 0x04D63A;
-            BossDropItemAddress = 0x180154;
-        }
-    }
-
-    public class SkullWoodsDungeon : Dungeon
-    {
-        public SkullWoodsDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Skull Woods";
-            DungeonType = DungeonType.SkullWoods;
-            DungeonCrystalTypeAddress = CrystalConstants.SkullWoodsCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.SkullWoodsCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 41;
-            BossAddress = 0x04D680;
-            BossDropItemAddress = 0x180155;
-        }
-    }
-
-    public class ThievesTownDungeon : Dungeon
-    {
-        public ThievesTownDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Thieves' Town";
-            DungeonType = DungeonType.ThievesTown;
-            DungeonCrystalTypeAddress = CrystalConstants.ThievesTownCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.ThievesTownCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 172;
-            BossAddress = 0x04D786;
-            BossDropItemAddress = 0x180156;
-        }
-    }
-
-    public class IcePalaceDungeon : Dungeon
-    {
-        public IcePalaceDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Ice Palace";
-            DungeonType = DungeonType.IcePalace;
-            DungeonCrystalTypeAddress = CrystalConstants.IcePalaceCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.IcePalaceCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 222;
-            BossAddress = 0x04D7EA;
-            BossDropItemAddress = 0x180157;
-        }
-    }
-
-    public class MiseryMireDungeon : Dungeon
-    {
-        public MiseryMireDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Misery Mire";
-            DungeonType = DungeonType.MiseryMire;
-            DungeonCrystalTypeAddress = CrystalConstants.MiseryMireCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.MiseryMireCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 144;
-            BossAddress = 0x04D74E;
-            BossDropItemAddress = 0x180158;
-        }
-    }
-
-    public class TurtleRockDungeon : Dungeon
-    {
-        public TurtleRockDungeon(int priority = 255) : base(priority)
-        {
-            Name = "Turtle Rock";
-            DungeonType = DungeonType.TurtleRock;
-            DungeonCrystalTypeAddress = CrystalConstants.TurtleRockCrystalTypeAddress;
-            DungeonCrystalAddress = CrystalConstants.TurtleRockCrystalAddress;
-            SelectedBoss = null;
-            BossRoomId = 164;
-            BossAddress = 0x04D776;
-            BossDropItemAddress = 0x180159;
-        }
-    }
-
-    public class GT1Dungeon : Dungeon
-    {
-        public GT1Dungeon(int priority = 255) : base(priority)
-        {
-            Name = "Ganon's Tower 1";
-            DungeonType = DungeonType.GanonsTower1;
-            DungeonCrystalTypeAddress = null;
-            DungeonCrystalAddress = null;
-            SelectedBoss = null;
-            BossRoomId = 28;
-            BossAddress = 0x04D666;
-            BossDropItemAddress = null;
-        }
-    }
-
-    public class GT2Dungeon : Dungeon
-    {
-        public GT2Dungeon(int priority = 255) : base(priority)
-        {
-            Name = "Ganon's Tower 2";
-            DungeonType = DungeonType.GanonsTower2;
-            DungeonCrystalTypeAddress = null;
-            DungeonCrystalAddress = null;
-            SelectedBoss = null;
-            BossRoomId = 108;
-            BossAddress = 0x04D706;
-            BossDropItemAddress = null;
-        }
-    }
-
-    public class GT3Dungeon : Dungeon
-    {
-        public GT3Dungeon(int priority = 255) : base(priority)
-        {
-            Name = "Ganon's Tower 3";
-            DungeonType = DungeonType.GanonsTower3;
-            DungeonCrystalTypeAddress = null;
-            DungeonCrystalAddress = null;
-            SelectedBoss = null;
-            BossRoomId = 77;
-            BossAddress = 0x04D6C8;
-            BossDropItemAddress = null;
-
-            DisallowedBosses.Add(BossType.Armos);
-            DisallowedBosses.Add(BossType.Arrghus);
-            DisallowedBosses.Add(BossType.Blind);
-            DisallowedBosses.Add(BossType.Lanmola);
-            DisallowedBosses.Add(BossType.Trixnexx);
-        }
-    }
 }
