+++ EnemizerLibrary/Dungeons/DungeonType.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EnemizerLibrary
+{
+    public enum DungeonType
+    {
+        EasternPalace,
+        DesertPalace,
+        TowerOfHera,
+        PalaceOfDarkness,
+        SwampPalace,
+        SkullWoods,
+        ThievesTown,
+        IcePalace,
+        MiseryMire,
+        TurtleRock,
+        GanonsTower1,
+        GanonsTower2,
+        GanonsTower3,
+        NotSet
+    }
+}
