+++ EnemizerLibrary/Dungeons/IcePalaceDungeon.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace EnemizerLibrary
+{
+    public class IcePalaceDungeon : Dungeon
+    {
+        public IcePalaceDungeon(int priority = 255) : base(priority)
+        {
+            Name = "Ice Palace";
+            DungeonType = DungeonType.IcePalace;
+            DungeonCrystalTypeAddress = CrystalConstants.IcePalaceCrystalTypeAddress;
+            DungeonCrystalAddress = CrystalConstants.IcePalaceCrystalAddress;
+            SelectedBoss = null;
+            BossRoomId = 222;
+            BossAddress = 0x04D7EA;
+            BossDropItemAddress = 0x180157;
+        }
+    }
+}
