+++ EnemizerLibrary/PotListing.cs
@@ -436,7 +436,11 @@ namespace EnemizerLibrary
                     itemPointer[1] = ROM_DATA[(0xDB67 + ((int)g.id * 2)) + 1];
                     int itemaddress = BitConverter.ToInt32(itemPointer, 0);
                     int addr = snestopc(itemaddress);
+
+                    // TODO: unused?
                     byte[] exportPots = new byte[roomPots.Count * 3];
+
+
                     for (int i = 0; i < roomPots.Count; i++)
                     {
                         ROM_DATA[addr + (i * 3) + 0] = roomPots[i].x;
